requirements.txt
codecov
coverage==6.3.2
Django==3.2.12
django-celery-results==2.3.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter
django-log-request-id==2.0.0
django-model-utils==4.2.0
django-redis==5.2.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
docker-compose==1.29.2
drf-yasg==1.20.0
factory-boy==3.2.1
feedparser==6.0.8
ipython==8.1.0
pre-commit==2.17.0
psycopg2==2.9.3
pymemcache==3.5.1
pytest==7.0.1
Redis==4.1.4
Sphinx==4.4.0
uWSGI==2.0.20
